refactor: delete network file declaration
