--- src/network-information-api.d.ts
@@ -1,25 +1,0 @@
-declare interface NavigatorNetworkInformation {
-  readonly connection?: NetworkInformation
-}
-type Megabit = number
-type Millisecond = number
-type EffectiveConnectionType = '2g' | '3g' | '4g' | 'slow-2g'
-type ConnectionType =
-  | 'bluetooth'
-  | 'cellular'
-  | 'ethernet'
-  | 'mixed'
-  | 'none'
-  | 'other'
-  | 'unknown'
-  | 'wifi'
-  | 'wimax'
-export interface NetworkInformation extends EventTarget {
-  readonly type?: ConnectionType
-  readonly effectiveType?: EffectiveConnectionType
-  readonly downlinkMax?: Megabit
-  readonly downlink?: Megabit
-  readonly rtt?: Millisecond
-  readonly saveData?: boolean
-  onchange?: EventListener
-}
