+++ tsconfig.json
@@ -20,5 +20,5 @@
     "jsx": "react-jsx"
   },
   "exclude": ["node_modules"],
-  "include": ["src"]
+  "include": ["./src"]
 }
